+++ core/models/model.py
@@ -172,7 +172,8 @@ class Model(nn.Module):
         if isinstance(self.embed_nodes, nn.Embedding):
             node_features = self.embed_nodes(self.g.ndata[GNN_NODE_LABELS_KEY])
         elif isinstance(self.embed_nodes, torch.Tensor):
-            node_features = self.embed_nodes[self.g.ndata[GNN_NODE_LABELS_KEY]]
+            # print(self.g.ndata[GNN_NODE_LABELS_KEY])
+            node_features = self.embed_nodes[self.g.ndata[GNN_NODE_LABELS_KEY].type(torch.long)]
         else:
             node_features = torch.zeros(self.g.number_of_nodes(), self.node_dim)
         node_features = node_features.cuda() if self.is_cuda else node_features
@@ -181,7 +182,7 @@ class Model(nn.Module):
         if isinstance(self.embed_edges, nn.Embedding):
             edge_features = self.embed_edges(self.g.edata[GNN_EDGE_LABELS_KEY])
         elif isinstance(self.embed_edges, torch.Tensor):
-            edge_features = self.embed_edges[self.g.edata[GNN_EDGE_LABELS_KEY]]
+            edge_features = self.embed_edges[self.g.edata[GNN_EDGE_LABELS_KEY].type(torch.long)]
         else:
             edge_features = None
